Making the fee into an integer rather than a double

+++ ICS372 GroupProject1/edu.metrostate.ics372.project1/edu/metrostate/ics372/project1/entities/Member.java
@@ -18,7 +18,7 @@ public class Member implements Serializable {
 	private Calendar date;
 
 	// Make the paid fee the same for each Member
-	private int fee = (int) 2.00;
+	private int fee = 2;
 
 	private static final String MEMBER_STRING = "M";
